+++ packages/twenty-front/src/modules/prefetch/components/PrefetchRunQueriesEffect.tsx
@@ -1,9 +1,18 @@
 import { useEffect } from 'react';
+import { useQuery } from '@apollo/client';
 import { useSetRecoilState } from 'recoil';
 
-import { usePrefetchFindManyCombinedQuery } from '@/prefetch/hooks/internal/usePrefetchCombinedQuery';
+import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
+import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { useUpsertFindManyRecordsQueryInCache } from '@/object-record/cache/hooks/useUpsertFindManyRecordsQueryInCache';
+import { useGenerateFindManyRecordsForMultipleMetadataItemsQuery } from '@/object-record/hooks/useGenerateFindManyRecordsForMultipleMetadataItemsQuery';
+import { useMapConnectionToRecords } from '@/object-record/hooks/useMapConnectionToRecords';
+import { MultiObjectRecordQueryResult } from '@/object-record/relation-picker/hooks/useMultiObjectRecordsQueryResultFormattedAsObjectRecordForSelectArray';
+import { ALL_FAVORITES_QUERY_KEY } from '@/prefetch/query-keys/AllFavoritesQueryKey';
+import { ALL_VIEWS_QUERY_KEY } from '@/prefetch/query-keys/AllViewsQueryKey';
 import { prefetchIsLoadedFamilyState } from '@/prefetch/states/prefetchIsLoadedFamilyState';
 import { PrefetchKey } from '@/prefetch/types/PrefetchKeys';
+import { isDefined } from '~/utils/isDefined';
 
 export const PrefetchRunQueriesEffect = () => {
   const setPrefetchAreViewsLoaded = useSetRecoilState(
@@ -13,14 +22,77 @@ export const PrefetchRunQueriesEffect = () => {
     prefetchIsLoadedFamilyState(PrefetchKey.AllFavorites),
   );
 
-  const { loading } = usePrefetchFindManyCombinedQuery();
+  const { objectMetadataItem: objectMetadataItemView } = useObjectMetadataItem({
+    objectNameSingular: CoreObjectNameSingular.View,
+  });
+
+  const { objectMetadataItem: objectMetadataItemFavorite } =
+    useObjectMetadataItem({
+      objectNameSingular: CoreObjectNameSingular.Favorite,
+    });
+
+  const prefetchFindManyQuery =
+    useGenerateFindManyRecordsForMultipleMetadataItemsQuery({
+      objectMetadataItems: [objectMetadataItemView, objectMetadataItemFavorite],
+      depth: 2,
+    });
+
+  if (!isDefined(prefetchFindManyQuery)) {
+    throw new Error('Could not prefetch recrds');
+  }
+
+  const mapConnectionToRecords = useMapConnectionToRecords();
+
+  const { data } = useQuery<MultiObjectRecordQueryResult>(
+    prefetchFindManyQuery,
+  );
+
+  const { upsertFindManyRecordsQueryInCache: upsertFindManyViewsInCache } =
+    useUpsertFindManyRecordsQueryInCache({
+      objectMetadataItem: objectMetadataItemView,
+    });
+
+  const { upsertFindManyRecordsQueryInCache: upsertFindManyFavoritesInCache } =
+    useUpsertFindManyRecordsQueryInCache({
+      objectMetadataItem: objectMetadataItemFavorite,
+    });
 
   useEffect(() => {
-    if (!loading) {
+    if (isDefined(data?.views)) {
+      upsertFindManyViewsInCache({
+        queryVariables: ALL_VIEWS_QUERY_KEY.variables,
+        depth: ALL_VIEWS_QUERY_KEY.depth,
+        objectRecordsToOverwrite:
+          mapConnectionToRecords({
+            objectRecordConnection: data.views,
+            objectNameSingular: CoreObjectNameSingular.View,
+            depth: 2,
+          }) ?? [],
+      });
       setPrefetchAreViewsLoaded(true);
+    }
+
+    if (isDefined(data?.views)) {
+      upsertFindManyFavoritesInCache({
+        queryVariables: ALL_FAVORITES_QUERY_KEY.variables,
+        depth: ALL_FAVORITES_QUERY_KEY.depth,
+        objectRecordsToOverwrite:
+          mapConnectionToRecords({
+            objectRecordConnection: data.favorites,
+            objectNameSingular: CoreObjectNameSingular.Favorite,
+            depth: 2,
+          }) ?? [],
+      });
       setPrefetchAreFavoritesLoaded(true);
     }
-  }, [setPrefetchAreViewsLoaded, loading, setPrefetchAreFavoritesLoaded]);
+  }, [
+    data,
+    setPrefetchAreViewsLoaded,
+    setPrefetchAreFavoritesLoaded,
+    upsertFindManyViewsInCache,
+    mapConnectionToRecords,
+    upsertFindManyFavoritesInCache,
+  ]);
 
   return <></>;
 };
